pyproject.toml
[project]
name = "medical-api"
version = "0.1.0"
description = ""
authors = [
    {name = "anapsantos1",email = "paulalimas17@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "django-filter (>=25.1,<26.0)"
]

[tool.poetry]
packages = [{include = "medical_api", from = "src"}]

[tool.poetry.group.dev.dependencies]
coverage = ">=7.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
